work version

+++ lab2/operations.cpp
@@ -3,7 +3,7 @@
 #include <omp.h>
 #include "operations.h"
 
-void PrintMat(double* mat, int size) {
+void PrintMat(float* mat, int size) {
     for (int i = 0; i < size; ++i) {
         for (int j = 0; j < size; ++j) {
             std::cout << mat[i * size + j] << "\t";
@@ -12,13 +12,13 @@ void PrintMat(double* mat, int size) {
     }
 }
 
-void PrintVect(double* vect, int size) {
+void PrintVect(float* vect, int size) {
     for (int i = 0; i < size; ++i) {
         std::cout << vect[i] << " ";
     }
 }
 
-void FillZero(double*& array, int size) {
+void FillZero(float*& array, int size) {
 
     for (int i = 0; i < size; ++i){
         array[i] = 0;
@@ -26,7 +26,7 @@ void FillZero(double*& array, int size) {
 
 }
 
-void FillMatrix(double*& matrix, int size) {
+void FillMatrix(float*& matrix, int size) {
 
     for (int i = 0; i < size; ++i) {
         for (int j = 0; j < size; ++j) {
@@ -36,7 +36,7 @@ void FillMatrix(double*& matrix, int size) {
 
 }
 
-void FillVectU(double*& u, int size) {
+void FillVectU(float*& u, int size) {
 
     for (int i = 0; i < size; ++i){
         u[i] = sin((2 * M_PI * i) / size);
@@ -44,9 +44,9 @@ void FillVectU(double*& u, int size) {
 }
 
 
-void FillVectB(double*& b, double*& u, double*& matrix, int size) {
+void FillVectB(float*& b, float*& u, float*& matrix, int size) {
     for (int i = 0; i < size; ++i) {
-        double temp = 0.0;
+        float temp = 0.0;
         for (int j = 0; j < size; ++j) {
             temp += matrix[i * size + j] * u[j];
         }
